enhancement(datadog): add APM Stats transform (#1095)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher quality. -->
This PR adds a new transform `apm_stats` that transforms `Trace` events into `TraceStat` events.
The proto file for `DDSketch` from `sketch-go` has also been added.


## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
Claude to copy the unit tests in the datadog-agent.

## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them here. -->

+++ lib/saluki-components/src/common/otlp/traces/transform.rs
@@ -231,8 +231,8 @@ pub fn otel_to_dd_span_minimal(
     let trace_id = convert_trace_id(&otel_span.trace_id);
     let span_id = convert_span_id(&otel_span.span_id);
     let parent_id = convert_span_id(&otel_span.parent_span_id);
-    let start = otel_span.start_time_unix_nano as i64;
-    let duration = (otel_span.end_time_unix_nano - otel_span.start_time_unix_nano) as i64;
+    let start = otel_span.start_time_unix_nano;
+    let duration = otel_span.end_time_unix_nano - otel_span.start_time_unix_nano;
     let mut meta: FastHashMap<MetaString, MetaString> = FastHashMap::default();
     meta.reserve(span_attributes.len() + resource_attributes.len());
     let mut metrics: FastHashMap<MetaString, f64> = FastHashMap::default();
